Correlogram: Extract code useful for Periodogram
